+++ apps/web/app/pages/report/[id].vue
@@ -118,6 +118,7 @@
                 :result="data.report"
                 :show-filename="false"
                 :show-pdf-ua-signals="false"
+                :link-manual-review="true"
               />
             </div>